[SQS] Initial commit
- Removed sqs library
- Substitution of queue string with env variables
